pyproject.toml
[tool.poetry]
name = "debias-bert"
version = "0.1.0"
description = ""
authors = ["Takuro Niitsuma <bakednt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
datasets = {git = "https://github.com/huggingface/datasets.git", rev = "master"}
transformers = "3.4"
comet_ml = "^3.2.2"
python-dotenv = "^0.14.0"
inflect = "^4.1.0"
tqdm = "4.49.0"
tensorboard = "^2.3.0"
torch = "^1.6.0"
pandas = "^1.1.2"
jupyterlab = "^2.2.8"
ipywidgets = "^7.5.1"
nltk = "^3.5"
pyjsg = "^0.10.0"
pytorch-lightning = "^1.0.1"
overrides = "^3.1.0"
catalyst = "^20.10.1"
allennlp = "1.2.1"
allennlp-models = "^1.2.0"
geomloss = "^0.2.3"
spacy = "^2.3.2"
scikit-learn = "^0.23.2"
polyglot = "^16.7.4"
PyICU = "^2.6"
Morfessor = "^2.0.6"
pycld2 = "^0.41"

[tool.poetry.dev-dependencies]
python-language-server = "^0.35.1"
black = "^20.8b1"
pyls-black = "^0.4.6"
ipdb = "^0.13.3"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
